+++ ChangeLog.md
@@ -11,6 +11,7 @@
     - TTP::RunnerCommand module is renamed RunnerVerb (todo #93)
     - Replace RunnerDaemon->startRun() with bootstrap() (todo #97), thus bumping minor candidate version number
     - ttp.pl push is now as verbose as ttp.pl pull (todo #91)
+    - alerts-monitor-daemon.pl: rename 'scanInterval' property with 'workerInterval'
 
 ### 4.5.1